initial commit
This commit gets all the essential features of the project in place.
The project is a C++ library that provides a hash function for strings.
The library is based on the CityHash library, which is a
high-performance hash function for strings. The library is designed to
be easy to use and efficient, and it is suitable for use in a wide range
of applications.

Signed-off-by: TLExpress <hank27161@gmail.com>